tweak: move presentational methods out of framework/engine
