+++ src/compiler/compiler-core/src/utils.ts
@@ -8,7 +8,7 @@ import {SourceMapConsumer} from 'source-map';
 import {Node, Project, TypeGuards} from 'ts-morph';
 import {findConfigFile, sys} from 'typescript';
 import {AcceptableType, AssemblerFactory, CompilerProvider, ComponentModule,
-  Constructor, DiezType, NamedComponentMap, PropertyDescription, TargetOutput} from './api';
+  Constructor, DiezType, NamedComponentMap, Presentable, PropertyDescription, TargetOutput} from './api';
 
 /**
  * A type guard for identifying a [[Constructor]] vs. a plain object.
@@ -418,4 +418,41 @@ export const setUpHandlebars = () => {
   registerPartial('comment', readFileSync(resolve(__dirname, '..', 'views', 'comment.handlebars')).toString());
   registerHelper('indent', indentContentHelper);
   registerHelper('ifIsCommentable', propertyIsCommentableHelper);
+};
+
+const isPrimitive = (value: any) => value === null || typeof value !== 'object';
+
+/**
+ * Checks if a value conforms to the [[Presentable]] interface
+ *
+ * @ignore
+ */
+export const isPresentable = (value: any): value is Presentable<any> => value && value.toPresentableValue instanceof Function;
+
+/**
+ * Presents a single property in a user-friendly way.
+ */
+export const presentProperty = <T>(value: T): any => {
+  if (isPresentable(value)) {
+    return presentProperty(value.toPresentableValue());
+  }
+
+  if (isPrimitive(value)) {
+    return String(value);
+  }
+
+  if (Array.isArray(value)) {
+    return `[${value.map(presentProperty)}]`;
+  }
+};
+
+/**
+ * Presents a record collection of properties in a user-friendly way.
+ */
+export const presentProperties = (value: Record<string, any>): Record<string, string> => {
+  const serialized: any = {};
+  for (const key in value) {
+    serialized[key] = presentProperty(value[key]);
+  }
+  return serialized;
 };
