+++ src/framework/engine/src/serialization.ts
@@ -1,14 +1,6 @@
-import {Presentable, Serializable} from './api';
+import {Serializable} from './api';
 
 const isSerializable = (value: any): value is Serializable<any> => value && value.serialize instanceof Function;
-
-/**
- * Checks if a value conforms to the [[Presentable]] interface
- *
- * @ignore
- */
-export const isPresentable = (value: any): value is Presentable<any> => value && value.toPresentableValue instanceof Function;
-
 const isPrimitive = (value: any) => value === null || typeof value !== 'object';
 
 /**
@@ -32,33 +24,6 @@ export const serialize = <T>(value: T): any => {
   const serialized: any = {};
   for (const key in value) {
     serialized[key] = serialize(value[key]);
-  }
-  return serialized;
-};
-
-export const presentProperty = <T>(value: T): any => {
-  if (isPresentable(value)) {
-    // Important! We must recursively serialize any subcomponents below.
-    return presentProperty(value.toPresentableValue());
-  }
-
-  if (isPrimitive(value)) {
-    return value;
-  }
-
-  if (Array.isArray(value)) {
-    return `[${value.map(presentProperty)}]`;
-  }
-};
-
-/**
- * An agnostic serializer for design token components, producing a stable and noncircular
- * representation of the data held in components.
- */
-export const presentProperties = <T>(value: T): any => {
-  const serialized: any = {};
-  for (const key in value) {
-    serialized[key] = presentProperty(value[key]);
   }
   return serialized;
 };
