+++ src/framework/stdlib/test/goldens/Bindings/diez-stdlib-test-stub-android/src/main/java/org/diez/stdlibTestStub/Point2D.kt
@@ -17,6 +17,7 @@ data class Point2D(
     /**
      * Point data.
      *
+     * 0
      */
     val x: Float,
     /**
