+++ src/framework/stdlib/test/goldens/Bindings/diez-stdlib-test-stub-android/src/main/java/org/diez/stdlibTestStub/Typograph.kt
@@ -42,6 +42,7 @@ data class Typograph(
     /**
      * Typograph data.
      *
+     * false
      */
     val shouldScale: Boolean,
     /**
@@ -53,6 +54,7 @@ data class Typograph(
     /**
      * Typograph data.
      *
+     * 0
      */
     val letterSpacing: Float,
     /**
