+++ src/framework/stdlib/test/goldens/Bindings/diez-stdlib-test-stub-ios/Sources/DiezStdlibTestStub/Components/Point2D.swift
@@ -8,6 +8,7 @@ public final class Point2D: NSObject, Decodable {
     /**
      * Point data.
      *
+     * 0
      */
     @objc public internal(set) var x: CGFloat
     /**
