+++ src/framework/stdlib/test/goldens/Bindings/diez-stdlib-test-stub-web/index.d.ts
@@ -190,6 +190,7 @@ export declare class Point2D {
   /**
    * Point data.
    *
+   * 0
    */
   x: number;
   /**
